refactor(shell): split lifecycle.rs into registry/open/io/reader/sharing
The single 1031-LOC shell/lifecycle.rs carried six unrelated concerns —
PTY spawn, input/resize/close, reader thread with UTF-8 codepoint-safe
splitting + OSC 0/1/2 title parsing, registry slot bookkeeping, and the
agent-facing share-mode / scrollback API. Split into one directory of
five focused submodules with `pub use` glob re-exports so every external
import path (`shell::ShellRegistry`, `shell::shell_open`, ...) stays
identical and `lib.rs` is untouched.

Layout:

  src-tauri/src/shell/lifecycle/
    mod.rs       — submodule decls + glob re-exports
    registry.rs  — ShellSlot, ShellRegistry, ChildHandle / ScrollbackHandle /
                   ShareHandle aliases, SCROLLBACK_BYTES, cfg(test) helpers
    open.rs      — shell_open, ShellOpenArgs, default_shell_{command,label}
    io.rs        — shell_input, shell_resize, shell_close (PTY control)
    reader.rs    — spawn_reader_thread, parse_osc_title, utf8_safe_split,
                   the three event payload types, READ_CHUNK_BYTES
    sharing.rs   — shell_set_share_mode, shell_read_scrollback, shell_write,
                   shell_list_shareable, ShareableShell, ScrollbackSnapshot,
                   ShellReadArgs, READ_DEFAULT_MAX, READ_HARD_CAP

The most invasive structural change is extracting the 90-line inline
reader-thread closure in shell_open into `reader::spawn_reader_thread`,
a free function generic over `Runtime`. open.rs now calls it as one
line; the UTF-8 carry / OSC scan / scrollback append / event-emit logic
travels with the parsing helpers that drive it. parse_osc_title was
already `pub(crate)` with zero external callers — tightened to private.

Visibility:

  - ShellRegistry stays pub (lib.rs `.manage(...)` consumer).
  - ShellSlot + fields, ChildHandle / ScrollbackHandle / ShareHandle,
    SCROLLBACK_BYTES are pub(super) so sibling submodules can construct
    and read slots without method ceremony.
  - parse_osc_title, utf8_safe_split, spawn_reader_thread,
    ShellOutputPayload / ShellExitPayload / ShellTitlePayload, READ_*
    constants are module-private.

Tests preserved 1:1, relocated to the submodule that owns the code
they exercise: 7 parse_osc_title + 8 utf8_safe_split tests in reader.rs;
5 write_keystrokes gating tests in sharing.rs; 5 registry-mechanic
tests (echo round-trip, resize, close-unknown, cleanup, concurrent
inserts) in registry.rs. Cross-submodule test setup goes through a
`registry::test_support` module exposing `open_raw` and `force_mode`
with `pub(in crate::shell::lifecycle)` visibility.

Zero behavior change. Verified with the full check gate: cargo check,
cargo clippy --all-targets -- -D warnings, cargo test --lib (158/158),
bunx vitest run (1124/1124).
